// ADDCOMP PM TOOL/Nove_Frontend/src/constants/enums.ts
export enum ERROR_MESSAGES {
  'NO_RECORD_FOUND' = 'Unfortunately, there are no records available at the moment.',
  'SERVER_ERROR' = 'Someting went to be wrong!',
  'POLICY_NOT_FOUND' = "Oops! We couldn't find any records at the moment. Please ensure that the provided information is accurate and try again.",
}

export enum MODULE_KEYS {
  LEADS = 'LEAD|LEAD|LATEST_LEAD_WITHOUT_OPPORTUNITY_STATUS',
  CRS_DRAWS = 'CRS DRAWS|CRS DRAWS|LATEST_CRS_DRAWS',
  NOC_CODES = 'NOC CODE|NOC CODE|LATEST_NOC_CODES',
  STUDY_PROGRAM = 'STUDY PROGRAM|STUDY PROGRAM|LATEST_STUDY_PROGRAM',

  COMPANY_INFORMATION = 'COMPANY INFORMATION|COMPANY INFORMATION|LATEST_COMPANY_INFORMATION',

  CUSTOMERS = 'CUSTOMER|CUSTOMER|LATEST_CUSTOMER_PR',
  CUSTOMER_SERVICES = "CUSTOMER|SERVICES|LATEST_CUSTOMER_SERVICE",
  CUSTOMER_INVOICE_AND_PAYMENT = "CUSTOMER|INVOICE AND PAYMENT|LATEST_CUSTOMER_INVOICE",
  CUSTOMER_PERSONAL_INFORMATION = "CUSTOMER|CUSTOMER PERSONAL INFORMATION|LATEST_CUSTOMER_PERSONAL_INFORMATION",
  CUSTOMER_RETAINER = "CUSTOMER|RETAINER|LATEST_RETAINER_LEADS",
  CUSTOMER_MANAGEMENT = "CUSTOMER|CUSTOMER MANAGEMENT|LATEST_WORKFLOW_BASIC",

  USERS = "USERS|USERS|LATEST_USER",
  USERS_CHANGE_ROLE = 'USERS|CHANGE ROLE|CHANGE_ROLE',

  SECURITY = 'SECURITY|SECURITY|ROLE_MODULE',
  ROLE_GROUP = "SECURITY|ROLE GROUP|LATEST_ROLE_GROUP",


  QUESTIONNAIRE = "QUESTIONNAIRE|QUESTIONNAIRE|LATEST_QUESTIONNAIRE",
  QUESTIONS = "QUESTIONNAIRE|QUESTIONS|LATEST_QUESTIONS",
  QUESTIONS_OPTIONS = "QUESTIONNAIRE|QUESTIONS OPTIONS|LATEST_QUESTIONS_OPTIONS",
  ANSWERS = "QUESTIONNAIRE|ANSWERS|LATEST_ANSWERS",

  CONVERSATION = "CONVERSATION|CONVERSATION|LATEST_CONVERSATION",
  CHAT_MESSAGES = "CONVERSATION|CHAT MESSAGES|LATEST_MESSAGES",
  CONVERSATION_PARTICIPANTS = "CONVERSATION|CONVERSATION PARTICIPANTS|LATEST_CONVERSATION_PARTICIPANTS",



  SERVICE = "SERVICE|SERVICE|LATEST_SERVICES",

  BRANCH = "DATA MANAGEMENT|BRANCH|LATEST_BRANCH",
  ZONE = "DATA MANAGEMENT|ZONE|ZONE",

  TASKBOARD = "TASKS|TASKBOARD|LATEST_TASKS",


  SETTING_CUSTOMER_AUTOMATION = "SETTING|WORKFLOW|LATEST_CUSTOMER_AUTOMATION",
  SETTING_TEMPLATE = "SETTING|TEMPLATE|LATEST_TEMPLATES",
  SETTING_AUTOMATION = "SETTING|CUSTOMER AUTOMATION|LATEST_WORKFLOW_BASIC"
}

export enum FILE_UPLOAD_KEYS {
  LEAD = 'LEAD',
  COMPANY_INFORMATION = 'COMPANY_INFORMATION',
  USERS = 'USERS',
  DOCUMENT_CHECKLIST = 'DOCUMENT_CHECKLIST',
}

export enum TEMPLATE_RELATED_MODULE_ID {
  LEAD = 'LEAD',
  G4SIGN = 'G4SIGN',
  TASK = 'TASK',
  QUOTES = 'QUOTES',
}